+++ src/fields/AssetConditionField.php
@@ -7,19 +7,7 @@ use craft\elements\conditions\assets\AssetCondition;
 class AssetConditionField extends ElementConditionField
 {
 
-	public string $conditionType = AssetCondition::class;
-	public ?string $elementType = Asset::class;
-
-	/*
-	 * Static
-	 */
-
-	/**
-	 * @inheritdoc
-	 */
-	public static function displayName(): string
-	{
-		return "Asset Condition";
-	}
+	protected static string $conditionType = AssetCondition::class;
+	protected static string $elementType = Asset::class;
 
 }