+++ src/fields/ElementConditionField.php
@@ -3,9 +3,11 @@ namespace TopShelfCraft\Conditions\fields;
 
 use Craft;
 use craft\base\conditions\ConditionInterface;
+use craft\base\Element;
 use craft\base\ElementInterface;
 use craft\base\Field;
 use craft\elements\conditions\ElementCondition;
+use craft\elements\conditions\ElementConditionInterface;
 use craft\helpers\Json;
 use yii\db\Schema;
 
@@ -16,13 +18,13 @@ class ElementConditionField extends Field
 	 * @var string The condition type for this field
 	 * @phpstan-var class-string<ConditionInterface>|null
 	 */
-	public string $conditionType = ElementCondition::class;
+	protected static string $conditionType = ElementCondition::class;
 
 	/**
-	 * @var string|null The element type to which the condition applies
+	 * @var string The element type to which the condition applies
 	 * @phpstan-var class-string<ElementInterface>|null
 	 */
-	public ?string $elementType = null;
+	protected static string $elementType = Element::class;
 
 	public bool $searchable = false;
 
@@ -31,18 +33,12 @@ class ElementConditionField extends Field
 		return Schema::TYPE_TEXT;
 	}
 
-	public function getElementConditionRuleType(): ?string
-	{
-		// TODO
-		return null;
-	}
-
 	protected function inputHtml(mixed $value, ?ElementInterface $element = null): string
 	{
 
-		$condition = $value instanceof ElementConditionFieldData
-			? $value->condition
-			: $this->normalizeValue($value, $element)->condition;
+		$condition = $value instanceof ElementConditionInterface
+			? $value
+			: $this->normalizeValue($value, $element);
 
 		$condition->mainTag = 'div';
 		$condition->name = $this->handle;
@@ -51,15 +47,13 @@ class ElementConditionField extends Field
 
 	}
 
-	public function normalizeValue(mixed $value, ?ElementInterface $element = null): ElementConditionFieldData
+	public function normalizeValue(mixed $value, ?ElementInterface $element = null): ElementConditionInterface
 	{
 
-		if ($value instanceof ElementConditionFieldData)
+		if ($value instanceof ElementConditionInterface)
 		{
 			return $value;
 		}
-
-		$fieldData = new ElementConditionFieldData();
 
 		if (is_string($value))
 		{
@@ -68,13 +62,11 @@ class ElementConditionField extends Field
 
 		$condition = is_array($value)
 			? Craft::$app->conditions->createCondition($value)
-			: Craft::createObject($this->conditionType);
-		$condition->elementType = $this->elementType;
+			: Craft::createObject(static::$conditionType);
+		$condition->elementType = static::$elementType;
 		$condition->sortable = true;
 
-		$fieldData->condition = $condition;
-
-		return $fieldData;
+		return $condition;
 
 	}
 
@@ -86,9 +78,9 @@ class ElementConditionField extends Field
 	public function serializeValue(mixed $value, ?ElementInterface $element = null): ?array
 	{
 
-		$condition = $value instanceof ElementConditionFieldData
-			? $value->condition
-			: $this->normalizeValue($value, $element)->condition;
+		$condition = $value instanceof ElementConditionInterface
+			? $value
+			: $this->normalizeValue($value, $element);
 
 		if (empty($condition->getConditionRules()))
 		{
@@ -105,12 +97,17 @@ class ElementConditionField extends Field
 
 	public static function displayName(): string
 	{
-		return "Element Condition";
+		return Craft::t('app', '{type} Condition', [
+			'type' => (static::$elementType)::displayName(),
+		]);
 	}
 
 	public static function valueType(): string
 	{
-		return ElementConditionFieldData::class .'|null';
+		return implode('|', [
+			static::$conditionType,
+			'null'
+		]);
 	}
 
 }