+++ src/fields/UserConditionField.php
@@ -7,19 +7,7 @@ use craft\elements\User;
 class UserConditionField extends ElementConditionField
 {
 
-	public string $conditionType = UserCondition::class;
-	public ?string $elementType = User::class;
-
-	/*
-	 * Static
-	 */
-
-	/**
-	 * @inheritdoc
-	 */
-	public static function displayName(): string
-	{
-		return "User Condition";
-	}
+	protected static string $conditionType = UserCondition::class;
+	protected static string $elementType = User::class;
 
 }